Split scoring.h: relocate popup components and TerminalResultState (refs #1194)
Continues the app/components/ audit (#1194) by splitting two non-score
concerns out of scoring.h to natural homes:

| Type                 | New home                      | Concern                            |
|---|---|---|
| ScorePopup           | components/popup.h (new)      | popup entity gameplay state        |
| PopupDisplay         | components/popup.h (new)      | popup entity render data           |
| TerminalResultState  | components/game_state.h       | terminal-phase "new best" snapshot |

ScorePopup and PopupDisplay are entity-bound popup-archetype components
(emplaced by spawn_score_popup, queried by popup_display_system /
ui_render_system / camera_system). They follow the existing
"one entity archetype per file" pattern (cf. obstacle.h, particle.h,
player.h) and now sit beside their owners.

TerminalResultState is a ctx singleton captured at game-over /
song-complete. It pairs with the GameOverState / DeathCause singletons
that PR #1231 moved to game_state.h, so it joins them there.

scoring.h keeps only ScoreState for now and re-includes popup.h /
game_state.h for source back-compat with the 18 direct includers (same
approach as rhythm.h after #1231). The ScoreState field-level cadence
split (accumulators vs animated display vs chain vs passive vs
high-score-current-song) is deferred to a follow-up cycle — that one
touches ~38 call sites and deserves its own focused PR.

Validation: 902 Catch2 tests / 2957 assertions pass; zero warnings on
both unity and non-unity Clang -Wall -Wextra -Werror builds.
